+++ Vaigachev/3/Warior.cpp
