--- Vaigachev/3/warior.h
@@ -1,17 +1,0 @@
-#pragma once
-#include "Object.h"
-#include <string.h>
-#include <stdio.h>
-
-class Warior :public Object
-{
-	int force;
-public:
-	Warior(ifstream& fin, const shared_ptr<Crown> t);
-	~Warior();
-	Warior* move(char option);
-	int give_force();
-	// thinkaboutittoo friend istream &operator>>(istream &in, Warior &ob);
-	friend ostream &operator<<(ostream &out, Warior &ob);
-};
-
